# openapi/v2/restful/template.yaml
openapi: 3.0.2
info:
  title: Horizon-Template-Restful
  description: Restful API for managing templates
  version: 2.0.0
servers:
  - url: 'http://localhost:8080/'
paths:
  /apis/core/v2/groups/{groupID}/templates:
    parameters:
      - name: groupID
        in: path
        description: id of group
        required: true
        schema:
          type: number
    post:
      tags:
        - template
      operationId: createTemplate
      summary: Create a template under specified group by group ID
      description: |
        Create a template under specified group by group ID. User must create a release when creating template, or the template would be useless.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                repository:
                  type: string
                token:
                  type: string
                release:
                  type: object
                  properties:
                    tag:
                      type: string
                    isRecommend:
                      type: boolean
                    description:
                      type: string

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"
...
  /apis/core/v2/templates/{template}:
    parameters:
      - name: template
        in: path
        description: id of template
        required: true
        schema:
          type: integer

    get:
      tags:
        - template
      operationId: getTemplate
      summary: Get the specified template
      description: |
        Get the specified template by template ID.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        description: id of template
                      name:
                        type: string
                        description: the name of template
                      description:
                        type: string
                        description: description of template
                      repository:
                        type: string
                        description: user-set gitlab url of tempalte
                      group:
                        type: integer
                        description: which group template belongs to
                      createBy:
                        type: integer
                        description: the creator of the template
                      updateBy:
                        type: integer
                        description: the user who updated the template
                        default:
                          description: Unexpected error
                          content:
                            application/json:
                              schema:
                                $ref: "common.yaml#/components/schemas/Error"
    put:
      tags:
        - template
      operationId: updateTemplate
      summary: Update the specified template
      description: |
        Update the specified template by template ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repository:
                  type: string
                  description: gitlab url of template repo
                token:
                  type: string
                  description: gitlab token to access the template repo
                description:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"

    delete:
      tags:
        - template
      operationId: deleteTemplate
      summary: Delete the specified template
      description: |
        Delete the specified template by template ID.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"

    post:
      tags:
        - release
      operationId: createRelease
      summary: Create release for specified template by template ID
      description: |
        Create release for specified template by template ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  type: string
                  description: tag in gitlab, it will also be release name
                isRecommend:
                  type: boolean
                description:
                  type: string

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"
...
  /apis/core/v2/templatereleases/{release}:
    parameters:
      - name: release
        in: path
        description: id of release
        required: true
        schema:
          type: number
    get:
      tags:
        - release
      operationId: getRelease
      summary: Get the specified release by release ID
      description: |
        Get the specified release by release ID.

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        description: the template release name
                      description:
                        type: string
                        description: the template release description
                      recommended:
                        type: boolean
                        description: is the most recommended release

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"

    put:
      tags:
        - release
      operationId: updateRelease
      summary: Get the specified release by release ID
      description: |
        Get the specified release by release ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isRecommend:
                  type: boolean
                description:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object


        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"

    delete:
      tags:
        - release
      operationId: deleteRelease
      summary: Delete the specified release by release ID
      description: |
        Delete the specified release by release ID.

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object


        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"
  /apis/core/v2/templatereleases/{release}/sync:
    parameters:
      - name: release
        in: path
        description: id of release
        required: true
        schema:
          type: number
    post:
      tags:
        - release
      operationId: syncReleaseToRepo
      summary: Upload the specified release to repo(such as harbor)
      description: |
        Upload the specified release to repo(such as harbor).
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "common.yaml#/components/schemas/Error"
...